+++ src/tfmindi/pp/__init__.py
@@ -1,5 +1,27 @@
 """Preprocessing functions for TF-MInDi."""
 
-from tfmindi.pp.seqlets import calculate_motif_similarity, create_seqlet_adata, extract_seqlets
+from tfmindi.pp.seqlets import (
+    calculate_motif_similarity,
+    create_seqlet_adata,
+    extract_seqlets,
+    get_example_contrib,
+    get_example_idx,
+    get_example_oh,
+    get_seqlet_matrices,
+    get_seqlet_matrix,
+    get_seqlet_oh,
+    get_seqlet_ohs,
+)
 
-__all__ = ["extract_seqlets", "calculate_motif_similarity", "create_seqlet_adata"]
+__all__ = [
+    "extract_seqlets",
+    "calculate_motif_similarity",
+    "create_seqlet_adata",
+    "get_example_idx",
+    "get_example_oh",
+    "get_example_contrib",
+    "get_seqlet_oh",
+    "get_seqlet_ohs",
+    "get_seqlet_matrix",
+    "get_seqlet_matrices",
+]